moved swing dependent utility to client package
