+++ core/src/main/java/net/adoptopenjdk/icedteaweb/client/policyeditor/PolicyEditor.java
@@ -109,7 +109,7 @@ import net.adoptopenjdk.icedteaweb.i18n.Translator;
 import net.adoptopenjdk.icedteaweb.ui.swing.SwingUtils;
 import net.sourceforge.jnlp.config.PathsAndFiles;
 import net.sourceforge.jnlp.runtime.JNLPRuntime;
-import net.sourceforge.jnlp.util.FileDialogFactory;
+import net.adoptopenjdk.icedteaweb.client.parts.dialogs.FileDialogFactory;
 import net.sourceforge.jnlp.util.FileUtils;
 import net.sourceforge.jnlp.util.FileUtils.OpenFileResult;
 import net.sourceforge.jnlp.util.ImageResources;
